+++ lib/str.ml
@@ -31,8 +31,8 @@ type regexp =
 
 let compile_regexp s c =
   let re = Emacs.re_no_emacs ~case:(not c) s in
-  { mtch = lazy (Compile.compile (Ast.seq [ Ast.start; re ]))
-  ; srch = lazy (Compile.compile re)
+  { mtch = lazy (Compile.compile_str (Ast.seq [ Ast.start; re ]))
+  ; srch = lazy (Compile.compile_str re)
   }
 ;;